requirements.txt
PyYAML>=6.0
flask>=3.1.2
streamlit>=1.28.0
requests>=2.28.0
python-dotenv>=1.0.0